Move createdAt to base entity
